feat: Initial packit setup
This commit contains the specfile for building the official package for Fedora
with a Packit setup.

Ultimately, a unified specfile is targeted for Fedora and any other rpm-based
distributions, e.g. openEuler.

And Packit(https://packit.dev/) is a tool for maintaining specfile within
upstream source. It requires a simple config file(.packit.yaml).

Log:
Signed-off-by: Robin Lee <cheeselee@fedoraproject.org>
Change-Id: I59b00d9a21b1c00a337684e5a4e3c9b2354d79f2
Reviewed-on: http://gerrit.uniontech.com/c/dtkgui/+/8522
Reviewed-by: zhangjide <zhangjide@uniontech.com>
Tested-by: zhangjide <zhangjide@uniontech.com>