fix: DCI icon paint failed when size is invalid
When the iconsize parameter does not exist, we still expect
to draw the DCI icon instead of returning an error. Because
the icon passed in by the user already contains the size of
the icon.

The ok variable is not used. We need to use it as the judgment
condition for the success of integer conversion.

Log:
Influence: DCI Icons
Change-Id: I3a26102819542718af97aa8ff373b05cfc5de61f
(cherry picked from commit 58088867b060426cee52a7656a0986862e6ef8c9)