feat: support group name on find dci icon file
Use the '/' to split the icon group name and it's actual icon name.
In the dtkdeclarative and dtkwidget, should be add the application
id for prefix for the icon name, eg: the "test" built-in icon in the
"org.deepin.app" application,  when to find it with
DIconTheme::findDciIconFile, the "iconName" argument should is
"org.deepin.app/test".

Influence: 无

Log:
Change-Id: Icc45faf16926993a6281aca5bee995b839c5b462