fix: Abnormal Dci Icon on high resolution screen
The AA_UseHighDpiPixmaps attribute is used to control the icon
blurring problem of the high-resolution screen. However, we do
not use this attribute, but expand the standard size of the icon
by default. This is unreasonable. Therefore, we modified the way
to enlarge the icon size to use AA_ After the usehighdpipixmaps
attribute, the icon blurring problem of the high-resolution screen
is solved.

Log:
Bug: https://pms.uniontech.com/bug-view-150309.html
Change-Id: Ie7d476cdec6c2ddea0a6432bc4a8a7d26aab5cac