fix: Don't change color if the HSL values no variation
If all the adjust parameters are 0, do not make any changes for the base
color. Because when trying to adjust the HSL value of an RGB spec color,
it will first convert the RGB spec to the HSL spec, and this conversion
will cause some precision loss, and cause the color has changed when
conversion back to the RGB spec.